Add option to serialize UTF-8 strings directly
The JSON serializer assumes that std::string and char* values are
encoded as Latin1, while cxxtools::String are unicode strings (using
characters from the BMP). This is impractical when serializing UTF-8
data. Instead of converting from UTF-8 to cxxtools::String and then using
the \uXXXX notation, add a flag inputUtf8 to tell the serializer that
std::string and char* values are UTF-8 encoded and can be inserted into
the UTF-8 JSON verbatim.

This is a backport of https://github.com/maekitalo/cxxtools/pull/17. The
UTF-8 codec in jsonserializer.cpp is disabled, to mimic the behavior of
9669f655f23c ("Optimize json serializer").